refactor: 重构插件模块并完善项目文档
将单体入口拆分为应用服务、领域逻辑、基础设施、插件入口、展示层和共享工具，并迁移静态资源、统一资源路径，降低模块耦合并保留 AstrBot 兼容入口。

补全源码实际使用的 HTTP、图片处理和数据处理依赖，增加模块结构与排行范围测试，同时清理无实现的占位目录和 Python 缓存。

依据当前代码、配置 schema 和指令实现重写 README，补充功能、安装、配置、使用示例、数据隐私、项目结构及贡献指南，移除旧架构和过时说明，便于用户正确部署和维护。
