+++ src/domain/ranking/push_scopes.py
